move logger under main src
